requirements.txt
flask==2.0.1
flask-cors==3.0.10
Pillow==9.5.0
transformers==4.30.2
torch==2.0.1